Split CodeGenerator to smaller files
